+++ examples/video_modes.rs
@@ -3,7 +3,13 @@ use winit::event_loop::EventLoop;
 fn main() {
     simple_logger::init().unwrap();
     let event_loop = EventLoop::new();
-    let monitor = event_loop.primary_monitor();
+    let monitor = match event_loop.primary_monitor() {
+        Some(monitor) => monitor,
+        None => {
+            println!("No primary monitor detected.");
+            return;
+        }
+    };
 
     println!("Listing available video modes:");